Lab_06/requirements.txt
requests~=2.25.0
Flask~=2.0.1
hazelcast-python-client
